Reorganize asset bucket upload
Move things around to make them a bit more flexible.
